Default logger can be configured via sysprops
Previously, the default logger was set statically; logd_logger for
Android builds, and stderr_logger for non-Android builds.

This change generalizes the stderr_logger into file_logger which makes
it possible to change the file where the logs are sent.  The file path
is from a new system property "ro.log.file_logger.path". If the system
property is not set, the destination file stream defaults to stderr for
backwards compatibility.

file_logger is enabled when the system property is set. This is useful
for variants of Android like Microdroid where logd doesn't exist and
programs are built with __ANDROID__.

Bug: 222592894
Test: Android still uses logd_logger.
Test: setprop ro.log.file_logger.path /dev/hvc2
      ... in Microdroid. Log is shown even without logd running.
Change-Id: I700cbfbbfc45e3ae58606fc0c2a77f592f7d652a