logd_integration_test: remove servicemanager
This was passing due to logs overflowing, but when they
are not overflowing, it will fail. The reason for this
is that before servicemanager initialized kernel
logging, the linker and libc logs in that process.

servicemanager actually does log quite a bit, but because
its logging to the kernel log, the PID is dropped there.

Bug: N/A
Test: atest logd_integration_test
Change-Id: I55d673efd99496e491559fff376b32434e517da4