liblog: Remove use of hardcoded PAGE_SIZE 4096
bionic hard codes the PAGE_SIZE macro as 4096. This is going away as
Android begins to support larger page sizes. Remove the usage of this
assumption from liblog source. The usage is not needed in this case
it's a secondary check on the read buffer size, which shouldn't matter
as the number of bytes read will account for when the read has finished.

Test: atest -c liblog-unit-tests
Bug: 295083738
Change-Id: I613304201ef78c7c99be99801ecd680b023e81a3
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>